// src/app/Models/Ev.ts
export class Ev {

    evId: number;
    evFiyat: number;
    evKat: string;
    evMetrekare: string;
    evOdaSayisi: string;
    evIsinma: string;
    evBanyoSayisi: string;
    evAdres: string;
    evEsya: string;
    evYas: string;
    evGorsel: string;
    evKategoriId: number;
    evKategoriAd: string;

}